Searchable Mineral Catalog V 1.2
The user can now search via the group search bar on the left side of the page. This search bar will disapear if the user is using a small screen. However, the user can still use the links at the bottom of the page to search. If the user is currently searching by group or letter the chosen item is highlighted.